backend/requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
google-generativeai
langchain